requirements.txt
psycopg[binary]
langchain
langgraph
langsmith
pgvector
sqlalchemy
fastapi
uvicorn
google-generativeai
python-dotenv 